add kadai-tasklist

+++ app/Http/Controllers/TasksController.php
@@ -17,13 +17,24 @@ class TasksController extends Controller
     // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        //タスク一覧を取得
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { //認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // タスクの一覧を取得
+            $tasks = Task::all();
+
+            // タスクの一覧を作成日時の降順で取得
+            $tasks = Task::orderBy('id', 'desc')->paginate(5);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+                ];
+        }
         
-        // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+        // Welcomeビューでそれらを標準
+        return view('welcome', $data);
     }
 
     /**
@@ -57,14 +68,14 @@ class TasksController extends Controller
             'content' => 'required|max:255',
             ]);
         
-        //タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        //認証済ユーザ（閲覧者）のタスクとして作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // 前のURLへリダイレクトさせる
+        return  redirect('/');
         
     }
 
@@ -77,13 +88,30 @@ class TasksController extends Controller
     //  getでtasks/(任意のid)にアクセスされた場合の「取得表示処理」
     public function show($id)
     {
+        // 認証済みユーザを取得
+        $user = \Auth::user();
+        
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
+        // 関係するモデルの件数をロード
+        $user->loadRelationshipCounts();
+        
+        // ユーザのタスク一覧を作成日時の降順で取得
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+        
+        
+          // 認証済ユーザ（閲覧社）がそのタスクの所有者である場合は、タスクを削除
+        if(\Auth::id() === $task->user_id) {
+        
         // タスク詳細ビューでそれを表示
         return view('tasks.show',[
+            'user'=>$user,
             'task'=>$task,
         ]);
+        }
+        // 前のURLへリダイレクトさせる
+        return back();
     }
 
     /**
@@ -122,10 +150,15 @@ class TasksController extends Controller
         
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+        
+        // 認証済ユーザ（閲覧社）がそのタスクの所有者である場合は、タスクを削除
+        if(\Auth::id() === $task->user_id) {
         // タスクを更新
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
+         
+        }
         
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -143,10 +176,11 @@ class TasksController extends Controller
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        // タスクを削除
-        $task->delete();
-        
-        // トップページへリダイレクトさせる
+        // 認証済ユーザ（閲覧社）がそのタスクの所有者である場合は、タスクを削除
+        if(\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        // 前のURLへリダイレクトさせる
         return redirect('/');
         
     }
